refactor(st-09087): deduplicate middleware controller wiring
