+++ planning/kanban-queue.md
@@ -1,11 +1,11 @@
 # Kanban Queue: AgentForge
 
-**Last Updated:** 2026-07-12
+**Last Updated:** 2026-07-13
 
 ## Queue Status Summary
 
-- **Ready:** 3 stories
-- **In Progress:** 0 stories
+- **Ready:** 2 stories
+- **In Progress:** 1 story
 - **In Review:** 0 stories
 - **Blocked:** 0 stories
 - **Backlog:** 4 stories
@@ -14,8 +14,6 @@
 
 ## Ready
 
-- `ST-09087` - DRY Middleware Rate-Limit and Concurrency Controllers
-  - Depends on: None
 - `ST-11001` - Publish Repository Security Boundary Policy
   - Depends on: None
 - `ST-11004` - Separate Worker Output from Supervisor Routing Input
@@ -31,7 +29,8 @@ None currently.
 
 ## In Progress
 
-None currently.
+- `ST-09087` - DRY Middleware Rate-Limit and Concurrency Controllers
+  - Depends on: None
 
 ## Blocked
